docs: add extra comments to ingest module

+++ src/sdk/media/transcode/util.py
@@ -2,13 +2,12 @@ import datetime
 from ffmpeg_streaming import FFProbe
 
 
-# TODO add tests
 def get_duration(input_file: str):
     """Get video time duration
 
     :param input_file: input path
     :return: (duration in seconds, timedelta hour)
-    :rtype: Union[int, datetime.timedelta]
+    :rtype: Union[float, datetime.timedelta]
     """
 
     ffprobe = FFProbe(input_file)
